testing masjid session max age
