+++ components/wrappers/auth.old.tsx
@@ -1,56 +1,76 @@
 'use client';
 
-import { useSession } from 'next-auth/react';
+import { useSession, signOut } from 'next-auth/react'; // Make sure signOut is imported
 import { useRouter } from 'next/navigation';
 import { Suspense, useEffect, useRef } from 'react';
 
 interface ChildrenType {
-  children: React.ReactNode;
+  children: React.ReactNode;
 }
 
 const FallbackLoading = () => (
-  <div style={{ 
-    display: 'flex', 
-    justifyContent: 'center', 
-    alignItems: 'center', 
-    height: '100vh', 
-    width: '100%', 
-    background: '#fff' 
-  }}>
-    <p>Loading...</p>
-  </div>
+  <div style={{ 
+    display: 'flex', 
+    justifyContent: 'center', 
+    alignItems: 'center', 
+    height: '100vh', 
+    width: '100%', 
+    background: '#fff' 
+  }}>
+    <p>Loading...</p>
+  </div>
 );
 
 const AuthProtectionWrapper = ({ children }: ChildrenType) => {
-  const { data: session, status } = useSession();
-  const router = useRouter();
-  const hasRedirected = useRef(false);
+  const { data: session, status } = useSession();
+  const router = useRouter(); // router is kept for potential future use, though signOut is preferred for auth redirects
+  const hasRedirected = useRef(false);
 
-  useEffect(() => {
-    // Only redirect once and when we're sure the user is unauthenticated
-    if (status === 'unauthenticated' && !hasRedirected.current) {
-      hasRedirected.current = true;
-      router.replace('/login'); // Use replace instead of push
+  useEffect(() => {
+    // If we are still checking the session or have already triggered a redirect, do nothing.
+    if (status === 'loading' || hasRedirected.current) {
+      return;
+    }
+
+    let shouldSignOut = false;
+
+    // --- Condition 1: The user is definitively unauthenticated. ---
+    if (status === 'unauthenticated') {
+        shouldSignOut = true;
+    }
+
+    // --- Condition 2 (Your New Check): Session exists, but the accessToken is missing. ---
+    // This indicates a malformed or incomplete session.
+    if (status === 'authenticated' && !session?.accessToken) {
+        console.error("Session is authenticated, but accessToken is missing. Signing out.");
+        shouldSignOut = true;
+    }
+
+    // If we need to sign out, we do it here.
+    if (shouldSignOut) {
+        hasRedirected.current = true;
+        // Using signOut is best because it clears the invalid session cookie.
+        router.replace('/login');
     }
-  }, [status, router]);
 
-  // Show loading while session is being determined
-  if (status === 'loading') {
-    return <FallbackLoading />;
-  }
+  }, [status, session]); // router can be removed from dependencies if only used for signOut
 
-  // If unauthenticated, show loading while redirecting
-  if (status === 'unauthenticated') {
-    return <FallbackLoading />;
-  }
+  // --- Rendering Logic ---
 
-  // Only render children if authenticated
-  if (status === 'authenticated' && session) {
-    return <Suspense fallback={<FallbackLoading />}>{children}</Suspense>;
-  }
+  // Show loading while the session status is being determined.
+  if (status === 'loading') {
+    return <FallbackLoading />;
+  }
 
-  // Fallback
-  return <FallbackLoading />;
+  // If the session is authenticated AND we have an accessToken, the user is valid.
+  if (status === 'authenticated' && session?.accessToken) {
+    return <Suspense fallback={<FallbackLoading />}>{children}</Suspense>;
+  }
+
+  // For all other cases (unauthenticated, session without accessToken),
+  // the useEffect is handling the redirect. We show a loading fallback
+  // to prevent a brief flash of content before the redirect occurs.
+  return <FallbackLoading />;
 };
 
 export default AuthProtectionWrapper;
