+++ lib/auth.ts
@@ -90,6 +90,7 @@ export const authOptions: NextAuthOptions = {
   ],
   session: {
     strategy: "jwt",
+    maxAge: 15 * 1,
   },
   callbacks: {
     async jwt({ token, user }) {
